+++ src/portalbackend/PortalBackend.PortalEntities/Entities/CompanyApplication.cs
@@ -53,11 +53,13 @@ public class CompanyApplication : IAuditableV1, IBaseEntity
     public Guid? ChecklistProcessId { get; set; }
 
     [AuditLastEditorV1]
-    public Guid? LastEditorId { get; set; }
+    public Guid? LastEditorId { get; private set; }
+
     // Navigation properties
     public virtual CompanyApplicationStatus? ApplicationStatus { get; set; }
     public virtual Company? Company { get; set; }
     public virtual Process? ChecklistProcess { get; set; }
     public virtual ICollection<Invitation> Invitations { get; private set; }
     public virtual ICollection<ApplicationChecklistEntry> ApplicationChecklistEntries { get; private set; }
+    public virtual Identity? LastEditor { get; private set; }
 }