+++ src/portalbackend/PortalBackend.PortalEntities/Entities/Notification.cs
@@ -66,5 +66,5 @@ public class Notification : IBaseEntity
     // Navigation properties
     public virtual CompanyUser? Receiver { get; set; }
     public virtual NotificationType? NotificationType { get; set; }
-    public virtual CompanyUser? Creator { get; set; }
+    public virtual Identity? Creator { get; set; }
 }