+++ tests/administration/Administration.Service.Tests/BusinessLogic/InvitationBusinessLogicTests.cs
@@ -24,7 +24,6 @@ using Org.Eclipse.TractusX.Portal.Backend.Framework.ErrorHandling;
 using Org.Eclipse.TractusX.Portal.Backend.Framework.Models.Configuration;
 using Org.Eclipse.TractusX.Portal.Backend.Mailing.SendMail;
 using Org.Eclipse.TractusX.Portal.Backend.PortalBackend.DBAccess;
-using Org.Eclipse.TractusX.Portal.Backend.PortalBackend.DBAccess.Models;
 using Org.Eclipse.TractusX.Portal.Backend.PortalBackend.DBAccess.Repositories;
 using Org.Eclipse.TractusX.Portal.Backend.PortalBackend.PortalEntities.Entities;
 using Org.Eclipse.TractusX.Portal.Backend.PortalBackend.PortalEntities.Enums;
@@ -50,7 +49,6 @@ public class InvitationBusinessLogicTests
     private readonly string _companyName;
     private readonly string _idpName;
     private readonly Guid _companyId;
-    private readonly IdentityData _identity;
     private readonly Guid _identityProviderId;
     private readonly Guid _applicationId;
     private readonly Func<UserCreationRoleDataIdpInfo, (Guid CompanyUserId, string UserName, string? Password, Exception? Error)> _processLine;
@@ -78,7 +76,6 @@ public class InvitationBusinessLogicTests
         _companyId = _fixture.Create<Guid>();
         _identityProviderId = _fixture.Create<Guid>();
         _applicationId = _fixture.Create<Guid>();
-        _identity = new(Guid.NewGuid().ToString(), Guid.NewGuid(), IdentityTypeId.COMPANY_USER, Guid.NewGuid());
 
         _processLine = A.Fake<Func<UserCreationRoleDataIdpInfo, (Guid CompanyUserId, string UserName, string? Password, Exception? Error)>>();
 
@@ -106,7 +103,7 @@ public class InvitationBusinessLogicTests
             _mailingService,
             _options);
 
-        await sut.ExecuteInvitation(invitationData, _identity.UserId).ConfigureAwait(false);
+        await sut.ExecuteInvitation(invitationData).ConfigureAwait(false);
 
         A.CallTo(() => _provisioningManager.GetNextCentralIdentityProviderNameAsync()).MustHaveHappened();
         A.CallTo(() => _provisioningManager.SetupSharedIdpAsync(A<string>.That.IsEqualTo(_idpName), A<string>.That.IsEqualTo(invitationData.organisationName), A<string?>._)).MustHaveHappened();
@@ -148,7 +145,7 @@ public class InvitationBusinessLogicTests
             _mailingService,
             _options);
 
-        Task Act() => sut.ExecuteInvitation(invitationData, _identity.UserId);
+        Task Act() => sut.ExecuteInvitation(invitationData);
 
         var error = await Assert.ThrowsAsync<ControllerArgumentException>(Act).ConfigureAwait(false);
         error.Message.Should().Be("email must not be empty (Parameter 'email')");
@@ -177,7 +174,7 @@ public class InvitationBusinessLogicTests
             _mailingService,
             _options);
 
-        Task Act() => sut.ExecuteInvitation(invitationData, _identity.UserId);
+        Task Act() => sut.ExecuteInvitation(invitationData);
 
         var error = await Assert.ThrowsAsync<ControllerArgumentException>(Act).ConfigureAwait(false);
         error.Message.Should().Be("organisationName must not be empty (Parameter 'organisationName')");
@@ -212,7 +209,7 @@ public class InvitationBusinessLogicTests
             _mailingService,
             _options);
 
-        Task Act() => sut.ExecuteInvitation(invitationData, _identity.UserId);
+        Task Act() => sut.ExecuteInvitation(invitationData);
 
         var error = await Assert.ThrowsAsync<TestException>(Act).ConfigureAwait(false);
         error.Message.Should().Be(_error.Message);
@@ -243,7 +240,7 @@ public class InvitationBusinessLogicTests
             _mailingService,
             _options);
 
-        Task Act() => sut.ExecuteInvitation(invitationData, _identity.UserId);
+        Task Act() => sut.ExecuteInvitation(invitationData);
 
         var error = await Assert.ThrowsAsync<TestException>(Act).ConfigureAwait(false);
         error.Message.Should().Be(_error.Message);
@@ -283,7 +280,7 @@ public class InvitationBusinessLogicTests
         A.CallTo(() => _provisioningManager.GetNextCentralIdentityProviderNameAsync()).Returns(_idpName);
 
         A.CallTo(() => _userProvisioningService.CreateOwnCompanyIdpUsersAsync(A<CompanyNameIdpAliasData>._, A<IAsyncEnumerable<UserCreationRoleDataIdpInfo>>._, A<CancellationToken>._))
-            .ReturnsLazily((CompanyNameIdpAliasData companyNameIdpAliasData, IAsyncEnumerable<UserCreationRoleDataIdpInfo> userCreationInfos, CancellationToken cancellationToken) =>
+            .ReturnsLazily((CompanyNameIdpAliasData _, IAsyncEnumerable<UserCreationRoleDataIdpInfo> userCreationInfos, CancellationToken _) =>
                 userCreationInfos.Select(userCreationInfo => _processLine(userCreationInfo)));
 
         A.CallTo(() => _processLine(A<UserCreationRoleDataIdpInfo>._)).ReturnsLazily(