+++ tests/externalsystems/Bpdm.Library/BpdmServiceTests.cs
@@ -112,23 +112,23 @@ public class BpdmServiceTests
     [Fact]
     public async Task FetchInputLegalEntity_WithValidResult_ReturnsExpected()
     {
+        // Arrange
         var externalId = _fixture.Create<string>();
         var data = _fixture.Build<BpdmLegalEntityOutputData>()
             .With(x => x.ExternalId, externalId)
             .With(x => x.Bpn, "TESTBPN")
-            .Create();
-        var pageOutputData = new PageOutputResponseBpdmLegalEntityData(
-            Enumerable.Repeat(data, 1),
-            Enumerable.Empty<BpdmErrorInfo>());
+            .CreateMany(1);
+        var pagination = new PageOutputResponseBpdmLegalEntityData(data);
+
         var options = new System.Text.Json.JsonSerializerOptions
         {
             PropertyNamingPolicy = System.Text.Json.JsonNamingPolicy.CamelCase,
             Converters = { new JsonStringEnumConverter(allowIntegerValues: false) }
         };
-        // Arrange
+
         var httpMessageHandlerMock = new HttpMessageHandlerMock(
             HttpStatusCode.OK,
-            pageOutputData.ToJsonContent(
+            pagination.ToJsonContent(
                 options,
                 "application/json")
             );
@@ -144,7 +144,7 @@ public class BpdmServiceTests
 
         // Assert
         result.Should().NotBeNull();
-        result.ExternalId.Should().Be(data.ExternalId);
+        result.ExternalId.Should().Be(externalId);
         result.Bpn.Should().Be("TESTBPN");
     }
 
@@ -193,39 +193,6 @@ public class BpdmServiceTests
         // Assert
         var ex = await Assert.ThrowsAsync<ServiceException>(Act).ConfigureAwait(false);
         ex.Message.Should().StartWith("Access to external system bpdm did not return a valid json response");
-        ex.IsRecoverable.Should().BeFalse();
-    }
-
-    [Fact]
-    public async Task FetchInputLegalEntity_WithError_ThrowsServiceException()
-    {
-        // Arrange
-        var error = new BpdmErrorInfo("SharingProcessError", "This is a test", "test");
-        var pageOutputData = new PageOutputResponseBpdmLegalEntityData(
-            Enumerable.Empty<BpdmLegalEntityOutputData>(),
-            Enumerable.Repeat(error, 1));
-        var options = new System.Text.Json.JsonSerializerOptions
-        {
-            PropertyNamingPolicy = System.Text.Json.JsonNamingPolicy.CamelCase,
-            Converters = { new JsonStringEnumConverter(allowIntegerValues: false) }
-        };
-        var httpMessageHandlerMock = new HttpMessageHandlerMock(
-            HttpStatusCode.OK,
-            pageOutputData.ToJsonContent(options, "application/json"));
-
-        var httpClient = new HttpClient(httpMessageHandlerMock)
-        {
-            BaseAddress = new Uri("https://base.address.com")
-        };
-        A.CallTo(() => _tokenService.GetAuthorizedClient<BpdmService>(_options.Value, A<CancellationToken>._)).Returns(httpClient);
-        var sut = new BpdmService(_tokenService, _options);
-
-        // Act
-        async Task Act() => await sut.FetchInputLegalEntity(_fixture.Create<string>(), CancellationToken.None).ConfigureAwait(false);
-
-        // Assert
-        var ex = await Assert.ThrowsAsync<ServiceException>(Act).ConfigureAwait(false);
-        ex.Message.Should().StartWith("The external system bpdm responded with errors ErrorCode: SharingProcessError, ErrorMessage: This is a test");
         ex.IsRecoverable.Should().BeFalse();
     }