+++ tests/shared/Tests.Shared/Extensions/ControllerExtensions.cs
@@ -22,7 +22,7 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Org.Eclipse.TractusX.Portal.Backend.Framework.ErrorHandling;
 using Org.Eclipse.TractusX.Portal.Backend.Framework.Models;
-using Org.Eclipse.TractusX.Portal.Backend.PortalBackend.DBAccess.Models;
+using Org.Eclipse.TractusX.Portal.Backend.PortalBackend.PortalEntities.Identities;
 using System.Security.Claims;
 
 namespace Org.Eclipse.TractusX.Portal.Backend.Tests.Shared.Extensions;
@@ -66,19 +66,15 @@ public static class ControllerExtensions
     /// Creates a claum for the identity user and adds it to the controller context
     /// </summary>
     /// <param name="controller">The controller that should be enriched</param>
-    /// <param name="iamUserId">Id of the iamUser</param>
     /// <param name="accessToken">Access token</param>
     /// <param name="identity">Identity of the user</param>
-    public static void AddControllerContextWithClaimAndBearer(this ControllerBase controller, string iamUserId, string accessToken, IdentityData? identity = null)
+    public static void AddControllerContextWithClaimAndBearer(this ControllerBase controller, string accessToken, IdentityData identity)
     {
         var claimsIdentity = new ClaimsIdentity();
-        claimsIdentity.AddClaims(new[] { new Claim(PortalClaimTypes.Sub, iamUserId) });
-        if (identity != null)
-        {
-            claimsIdentity.AddClaims(new[] { new Claim(PortalClaimTypes.IdentityId, identity.UserId.ToString()) });
-            claimsIdentity.AddClaims(new[] { new Claim(PortalClaimTypes.IdentityType, Enum.GetName(identity.IdentityType) ?? throw new UnexpectedConditionException("itentityType should never be out of enum-values range here")) });
-            claimsIdentity.AddClaims(new[] { new Claim(PortalClaimTypes.CompanyId, identity.CompanyId.ToString()) });
-        }
+        claimsIdentity.AddClaims(new[] { new Claim(PortalClaimTypes.Sub, identity.UserEntityId) });
+        claimsIdentity.AddClaims(new[] { new Claim(PortalClaimTypes.IdentityId, identity.UserId.ToString()) });
+        claimsIdentity.AddClaims(new[] { new Claim(PortalClaimTypes.IdentityType, Enum.GetName(identity.IdentityType) ?? throw new UnexpectedConditionException("itentityType should never be out of enum-values range here")) });
+        claimsIdentity.AddClaims(new[] { new Claim(PortalClaimTypes.CompanyId, identity.CompanyId.ToString()) });
 
         var httpContext = new DefaultHttpContext
         {